moved functions from helpers to model
